Starting fresh again
I realized that since I didn't have enough of a grasp of the problem that I am just going around blind. What I am going to do now is simply the problem in a way that is easy to see and change in order to really learn the systems that are causing this bug.